+++ kinalcoffeeshop/src/main/java/org/asco_devs/kinalcoffeeshop/dominio/dto/ModPedidoDto.java
@@ -3,7 +3,7 @@ package org.asco_devs.kinalcoffeeshop.dominio.dto;
 import java.math.BigDecimal;
 
 public record ModPedidoDto(
-        BigDecimal total,
+        Double total,
         String state,
         Long studentId,
         Long creditUserId
